authentication af endpoints + auth features smidt i seperat fil, ved brug af Blueprints + login required decorator
