+++ hello_app/templates/layout.html
@@ -33,10 +33,10 @@
         {% if session['user'] %}
         <!-- Logged in as text -->
         <span class="navbar-item">Logged in as: <span style="color:orange;font-weight:bold;">{{ session['user'] }}</span></span>
-        <a href="{{ url_for('logout') }}" class="navbar-item">Logout</a>
+        <a href="{{ url_for('auth.logout') }}" class="navbar-item">Logout</a>
         {% else %}
-      <a href="{{ url_for('login') }}" class="navbar-item">Login</a>
-      <a href="{{ url_for('create_account') }}" class="navbar-item">Signup</a>
+      <a href="{{ url_for('auth.login') }}" class="navbar-item">Login</a>
+      <a href="{{ url_for('auth.create_account') }}" class="navbar-item">Signup</a>
       {% endif %}
       </div>
     </div>
