+++ hello_app/templates/login.html
@@ -15,7 +15,7 @@
             </div>
         {% endif %}
         {% endwith %}
-        <form method="POST" action="/login">
+        <form method="POST" action="/auth/login">
             <div class="field">
                 <div class="control">
                     <input class="input is-large" type="text" name="username" placeholder="Your Username" autofocus="">
