+++ hello_app/templates/signup.html
@@ -15,7 +15,7 @@
             </div>
         {% endif %}
         {% endwith %}
-        <form method="POST" action="/signup">
+        <form method="POST" action="/auth/signup">
             <div class="field">
                 <div class="control">
                     <input class="input is-large" type="email" name="email" placeholder="Email" autofocus="" value="Test@test.com">
