+++ hello_app/views.py
@@ -6,11 +6,9 @@ from flask import Flask, render_template, request, redirect, url_for, flash, ses
 from . import app
 from .models import Note, Account
 from .utils import *
+from .auth import auth, login_required
 
-
-
-
-
+app.register_blueprint(auth, url_prefix='/auth')
 
 # Vores landing Page - Der viser listen over noter
 @app.route("/")
@@ -19,8 +17,8 @@ def home():
 
 
 # Create note
-#Login required (what if you're already logged in)
 @app.route("/create/", methods=["GET","POST"])
+@login_required
 def create_note():
 
     if request.method == "GET":
@@ -32,8 +30,8 @@ def create_note():
 
     return redirect(url_for('home'))
 
-#Login required (what if you're already logged in =session)
 @app.route("/edit/<id>", methods=["GET","POST"])
+@login_required
 def edit(id = None):
 
     note = find_note(id)
@@ -67,6 +65,7 @@ def view(id = None):
 
 
 @app.route("/delete/<id>")
+@login_required
 def delete_note(id = None):
     
     if id is None:
@@ -96,67 +95,3 @@ def search_results():
     # Søg efter noter, der matcher søgeordet
     search_results = searchbar(query)
     return render_template('search_results.html', results=search_results, query=query)
-
-@app.route("/signup/", methods=["GET","POST"])
-def create_account():
-    if request.method == "POST":
-        # Create account post logik
-  
-        if not request.form['username']:
-            flash('Username is required', 'error')
-            return redirect(url_for('create_account'))
-        if not request.form['password']:
-            flash('Password is required', 'error')
-            return redirect(url_for('create_account'))
-        if not request.form['email']:
-            flash('Email is required', 'error')
-            return redirect(url_for('create_account'))
-        
-        # Check if username already exists
-        if dbsession.query(Account).filter_by(username=request.form['username']).first() is not None:
-            flash('Username already exists', 'error')
-            return redirect(url_for('create_account'))
-        
-        try:
-            dbsession.add(Account(username = request.form['username'], password = request.form['password'], email = request.form['email']))
-            dbsession.commit()
-            flash('Account created successfully!', 'success')
-            
-        except:
-            flash('Error creating account', 'error')
-            return redirect(url_for('create_account'))
-        return redirect(url_for('home'))
-        
-    else:
-        return render_template("signup.html")
-
-#login route
-@app.route('/login', methods=["GET", "POST"])
-def login():
-    if request.method == "POST":
-
-        username = request.form.get('username')
-        password = request.form.get('password')
-
-        # Query the database for the account with provided username and password
-        account = dbsession.query(Account).filter_by(username=username, password=password).first()
-
-         
-        if account:   
-
-            session['user'] = account.username
-
-            flash(f'Login successful for {account.username}', 'success')
-            return redirect(url_for('home'))
-        else:
-            flash('Invalid username or password', 'error')
-            return redirect(url_for('login'))
-
-    # If request method is GET, render the login template
-    return render_template('login.html')
-@app.route('/logout')
-def logout():
-    user_logout = session['user']
-    session.pop('user', None) 
-    flash(f'You have been logged out, {user_logout}', 'success')
-    return redirect(url_for('home'))
